+++ lib/authentication/auth_workload_identity/auth_workload_identity.ts
@@ -7,7 +7,7 @@ import Logger from '../../logger';
 import { getAzureAttestationToken } from './attestation_azure';
 import { getGcpAttestationToken } from './attestation_gcp';
 import AuthenticationTypes from '../authentication_types';
-import { validateAccessUrl } from './host_allowlist';
+import { validateHost } from './host_allowlist';
 
 class AuthWorkloadIdentity implements AuthClass {
   tokenProvider!: WorkloadIdentityProviderKey;
@@ -25,11 +25,11 @@ class AuthWorkloadIdentity implements AuthClass {
     const {
       workloadIdentityProvider: provider,
       workloadIdentityImpersonationPath: impersonationPath,
-      accessUrl,
+      host,
     } = this.connectionConfig;
     let token: string;
 
-    validateAccessUrl(accessUrl);
+    validateHost(host);
 
     if (
       impersonationPath &&